Automatic leaderboard upload at the end of I210 training

+++ flow/visualize/i210_replay.py
@@ -6,6 +6,7 @@ from copy import deepcopy
 import numpy as np
 import json
 import os
+import os.path
 import pytz
 import subprocess
 import time
@@ -241,8 +242,8 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
     metadata['submission_time'].append(cur_time)
     metadata['network'].append(network_name_translate(env.network.name.split('_20')[0]))
     metadata['is_baseline'].append(str(args.is_baseline))
-    if args.to_aws:
-        name, strategy = get_configuration()
+    if args.use_s3:
+        name, strategy = get_configuration(args.submitter_name, args.strategy_name)
         metadata['submitter_name'].append(name)
         metadata['strategy'].append(strategy)
 
@@ -362,8 +363,12 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
                 '{0}/test_time_rollout/{1}'.format(dir_path, emission_filename)
 
             output_path = os.path.join(output_dir, '{}-emission.csv'.format(exp_name))
-            # convert the emission file into a csv file
-            emission_to_csv(emission_path, output_path=output_path)
+            if os.path.exists(emission_path.replace('emission.xml', '0_emission.csv')):
+                # csv already exists
+                os.rename(emission_path.replace('emission.xml', '0_emission.csv'), output_path)
+            else:
+                # convert the emission file into a csv file
+                emission_to_csv(emission_path, output_path=output_path)
 
             # generate the trajectory output file
             trajectory_table_path = os.path.join(dir_path, '{}.csv'.format(source_id))
@@ -384,7 +389,8 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
             print("\nGenerated emission file at " + output_path)
 
             # delete the .xml version of the emission file
-            os.remove(emission_path)
+            if os.path.exists(emission_path):
+                os.remove(emission_path)
 
         all_trip_energies = os.path.join(output_dir, '{}-all_trip_energies.npy'.format(exp_name))
         np.save(all_trip_energies, dict(all_trip_energy_distribution))
@@ -500,15 +506,19 @@ def create_parser():
         action='store_true',
         help='specifies whether this is a baseline run'
     )
+    parser.add_argument('--submitter_name', type=str, required=False, default=None,
+                        help='Name of the submitter (replaces the prompt asking for '
+                             'the name) (stored locally so only necessary once)')
+    parser.add_argument('--strategy_name', type=str, required=False, default=None,
+                        help='Name of the training strategy (replaces the prompt '
+                        'asking for the strategy)')
     return parser
 
 
-if __name__ == '__main__':
+def generate_graphs(args):
+    """Generate the graphs."""
     date = datetime.now(tz=pytz.utc)
     date = date.astimezone(pytz.timezone('US/Pacific')).strftime("%m-%d-%Y")
-
-    parser = create_parser()
-    args = parser.parse_args()
 
     rllib_config = None
     rllib_result_dir = None
@@ -520,12 +530,13 @@ if __name__ == '__main__':
 
     flow_params = deepcopy(I210_MA_DEFAULT_FLOW_PARAMS)
 
-    if args.multi_node:
-        ray.init(redis_address='localhost:6379')
-    elif args.local:
-        ray.init(local_mode=True, object_store_memory=200 * 1024 * 1024)
-    else:
-        ray.init(num_cpus=args.num_cpus + 1, object_store_memory=200 * 1024 * 1024)
+    if not ray.is_initialized():
+        if args.multi_node:
+            ray.init(redis_address='localhost:6379')
+        elif args.local:
+            ray.init(local_mode=True, object_store_memory=200 * 1024 * 1024)
+        else:
+            ray.init(num_cpus=args.num_cpus + 1, object_store_memory=200 * 1024 * 1024)
 
     if args.exp_title:
         output_dir = os.path.join(args.output_dir, args.exp_title)
@@ -573,3 +584,10 @@ if __name__ == '__main__':
                 p1.wait(50)
             except Exception as e:
                 print('This is the error ', e)
+
+
+if __name__ == '__main__':
+    parser = create_parser()
+    args = parser.parse_args()
+
+    generate_graphs(args)
